Separate tests for dictionaryTypes into smaller files
