+++ apps/server/test/integration/fixtures/createDictionary.json
@@ -1,5 +1,5 @@
 {
-	"name": "Test Dictionary",
+	"name": "test_dictionary",
 	"version": "1.2",
 	"schemas": [
 		{
