+++ apps/server/test/integration/fixtures/createKeyValue.json
@@ -1,5 +1,5 @@
 {
-	"name": "Key Value Meta",
+	"name": "key_value_meta",
 	"version": "1.0",
 	"references": {
 		"enums": {
