+++ libraries/dictionary/test/types/references.spec.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -18,26 +18,26 @@
  */
 
 import { expect } from 'chai';
-import { replaceReferences } from '../src/references';
+import { replaceReferences } from '../../src/references';
 
-import noReferencesSectionInput from './fixtures/references/no_references_section/input';
-import noReferencesSectionOutput from './fixtures/references/no_references_section/output';
-import emptyReferencesInput from './fixtures/references/empty_references_section/input';
-import emptyReferencesOutput from './fixtures/references/empty_references_section/output';
-import simpleReferencesInput from './fixtures/references/simple_references/input';
-import simpleReferencesOutput from './fixtures/references/simple_references/output';
-import codeListReferencesInput from './fixtures/references/codeList_references/input';
-import codeListReferencesOutput from './fixtures/references/codeList_references/output';
-import referencesWithinReferencesInput from './fixtures/references/references_within_references/input';
-import referencesWithinReferencesOutput from './fixtures/references/references_within_references/output';
-import scriptReferencesInput from './fixtures/references/script_references/input';
-import scriptReferencesOutput from './fixtures/references/script_references/output';
-import regexReferencesInput from './fixtures/references/regex_reference/input';
-import regexReferencesOutput from './fixtures/references/regex_reference/output';
-import regexArrayReferencesInput from './fixtures/references/regex_reference/input_with_array';
-import nonExistingReferencesInput from './fixtures/references/non_existing_references/input';
-import cyclicReferencesInput from './fixtures/references/cyclic_references/input';
-import selfReferencesInput from './fixtures/references/self_references/input';
+import noReferencesSectionInput from '../fixtures/references/no_references_section/input';
+import noReferencesSectionOutput from '../fixtures/references/no_references_section/output';
+import emptyReferencesInput from '../fixtures/references/empty_references_section/input';
+import emptyReferencesOutput from '../fixtures/references/empty_references_section/output';
+import simpleReferencesInput from '../fixtures/references/simple_references/input';
+import simpleReferencesOutput from '../fixtures/references/simple_references/output';
+import codeListReferencesInput from '../fixtures/references/codeList_references/input';
+import codeListReferencesOutput from '../fixtures/references/codeList_references/output';
+import referencesWithinReferencesInput from '../fixtures/references/references_within_references/input';
+import referencesWithinReferencesOutput from '../fixtures/references/references_within_references/output';
+import scriptReferencesInput from '../fixtures/references/script_references/input';
+import scriptReferencesOutput from '../fixtures/references/script_references/output';
+import regexReferencesInput from '../fixtures/references/regex_reference/input';
+import regexReferencesOutput from '../fixtures/references/regex_reference/output';
+import regexArrayReferencesInput from '../fixtures/references/regex_reference/input_with_array';
+import nonExistingReferencesInput from '../fixtures/references/non_existing_references/input';
+import cyclicReferencesInput from '../fixtures/references/cyclic_references/input';
+import selfReferencesInput from '../fixtures/references/self_references/input';
 
 describe('Replace References', () => {
 	it('Should return the same original schema if dictionary does not contain a references section', () => {
